move localStorage to utils dir
this will be needed by CosmWasm logic, but the rest of the NEAR stuff
won't
